requirements.txt
alembic==1.12.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
mailjet-rest==1.3.4
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2==2.9.7
pycoingecko==3.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.3.7
WTForms==3.0.1
